+++ backend/src/main/java/com/example/demo/global/security/SecurityConfig.java
@@ -8,8 +8,8 @@ import com.example.demo.global.security.handler.JwtAuthenticationSuccessHandler;
 import com.example.demo.global.jwt.service.TokenService;
 import com.example.demo.global.response.RsData;
 import com.example.demo.global.security.filter.LoginAuthenticationFilter;
-import com.example.demo.global.security.handler.CustomAccessDeniedHandler;
-import com.example.demo.global.security.handler.CustomAuthenticationEntryPoint;
+import com.example.demo.global.security.exception.CustomAccessDeniedHandler;
+import com.example.demo.global.security.exception.CustomAuthenticationEntryPoint;
 import com.example.demo.global.security.handler.CustomLogoutHandler;
 import com.example.demo.global.util.Ut;
 import jakarta.servlet.http.HttpServletResponse;
@@ -28,8 +28,6 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.annotation.web.configurers.HeadersConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.cors.CorsConfiguration;
