+++ backend/src/main/java/com/example/demo/global/security/exception/CustomAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package com.example.demo.global.security.handler;
+package com.example.demo.global.security.exception;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
